requirements.txt
aiohttp
lxml
img2pdf
Pillow
telethon
